+++ ServerFramework/PassiveMode.cpp
@@ -19,7 +19,7 @@ void PassiveMode::run()
 		_broadcasters[engineType]->broadcast(evt);
 
 		if (evt.type == E_EngineEventType::EVENT_NET_RECV)
-			MemoryBlockPool::release(evt.blockSize, evt.data);
+			MemoryBlockPool::release((int)evt.blockSize, evt.data);
 	}
 }
 #pragma endregion
